Added: Save to most recently opened database Added: Edit/delete entries in the currently loaded database, including a fuzzy search to help you find the right entry.
Further potential improvements have been added to vtBugsAndFeatures.txt, but these may not be implements for quite a while as a) they're minor, b) they don't prevent intended use or functionality of the program and c) this has been primarily a learning excercise and I am eager to turn my attentions to a more modern implementation of the program.

If anyone is desperate to further this version of the vocab tester, contact me through github.